Add files via upload
Before contacting me with questions, please read the paper at https://doi.org/10.1101/322511 to see if your questions are answered there. In the case of any confusion, the paper should be considered the more authoritative source. The python files are meant to read a BLAST output file, and the data analysis is then performed in R.